chore: clean up naming
