+++ src/entities/fungible-token/index.ts
@@ -1,4 +1,4 @@
-import { fungibleTokensBalancesRender } from "./ui/balances";
+import { fungibleTokensBalancesRender } from "./ui/fungible-token";
 
 export class FungibleToken {
     static allBalancesRender = fungibleTokensBalancesRender;
