+++ src/entities/fungible-token/model/fungible-token.ts
