+++ src/entities/fungible-token/ui/fungible-token.tsx
@@ -1,7 +1,7 @@
 import { TokenLabel } from "../../../shared/ui/components";
 import type { DaoContracts } from "../../types";
 import { FungibleTokenFormat } from "../lib/format";
-import { FungibleTokenBalancesModel } from "../model/balances";
+import { FungibleTokenBalancesModel } from "../model/fungible-token";
 
 interface FungibleTokensBalancesRenderProps {
     daoContracts: DaoContracts;
