+++ src/entities/native-token/index.ts
@@ -1,4 +1,4 @@
-import { nativeTokenBalancesRender } from "./ui/balances";
+import { nativeTokenBalancesRender } from "./ui/native-token";
 
 export class NativeToken {
     static balancesRender = nativeTokenBalancesRender;
