+++ src/entities/native-token/model/native-token.ts
