+++ src/entities/task/index.ts
@@ -1,5 +1,1 @@
-import { TaskCard } from "./ui/card";
-
-export const Task = {
-    Card: TaskCard,
-};
+export { Task } from "./ui/task";
