+++ src/entities/task/ui/task.jsx
@@ -2,10 +2,10 @@ import React, { Component } from "react";
 import { Draggable } from "react-beautiful-dnd";
 
 import { Family } from "../../../families/index.js";
-import "./card.scss";
+import "./task.scss";
 
 /* TODO: Decompose model and view */
-export class TaskCard extends Component {
+export class Task extends Component {
     id;
 
     constructor(props) {
