+++ src/pages/app/app.tsx
@@ -1,0 +1,11 @@
+import { Layout } from "../../widgets";
+
+export const AppPage = () => {
+    window.PAGE = "app";
+
+    return (
+        <>
+            <Layout />
+        </>
+    );
+};
