+++ src/pages/app/index.tsx
@@ -1,11 +1,1 @@
-import { Layout } from "../../widgets";
-
-export const AppPage = () => {
-    window.PAGE = "app";
-
-    return (
-        <>
-            <Layout />
-        </>
-    );
-};
+export { AppPage } from "./app";
