+++ src/shared/ui/components/dialog/index.ts
@@ -1,0 +1,1 @@
+export { Dialog } from "./dialog";
