+++ src/shared/ui/components/facet/facet.scss
@@ -1,7 +1,7 @@
 @use "sass/size";
 @use "sass/color";
 
-.Card {
+.Facet {
     display: flex;
     flex-flow: column nowrap;
     position: relative;
