+++ src/shared/ui/components/facet/index.ts
@@ -1,0 +1,1 @@
+export { Facet } from "./facet";
