+++ src/shared/ui/components/icons/near/near.tsx
