+++ src/shared/ui/components/index.ts
@@ -1,9 +1,9 @@
-export { Card } from "./card";
-export { Dialog } from "./dialog";
+export { Facet } from "./facet/facet";
+export { Dialog } from "./dialog/dialog";
 export * from "./icons";
-export { PopupMenu } from "./popup-menu";
-export { Scrollable } from "./scrollable";
-export { Table } from "./table";
-export { Tabs } from "./tabs";
-export { TokenLabel } from "./token-label";
-export { Tooltip } from "./tooltip";
+export { PopupMenu } from "./popup-menu/popup-menu";
+export { Scrollable } from "./scrollable/scrollable";
+export { Table } from "./table/table";
+export { Tabs } from "./tabs/tabs";
+export { TokenLabel } from "./token-label/token-label";
+export { Tooltip } from "./tooltip/tooltip";
