+++ src/shared/ui/components/popup-menu/index.ts
@@ -1,0 +1,1 @@
+export { PopupMenu } from "./popup-menu";
