+++ src/shared/ui/components/scrollable/index.ts
@@ -1,0 +1,1 @@
+export { Scrollable } from "./scrollable";
