+++ src/shared/ui/components/table/index.ts
@@ -1,0 +1,1 @@
+export { Table } from "./table";
