+++ src/shared/ui/components/tabs/index.ts
@@ -1,0 +1,1 @@
+export { Tabs } from "./tabs";
