+++ src/shared/ui/components/tabs/tabs.tsx
