+++ src/shared/ui/components/token-label/index.ts
@@ -1,0 +1,1 @@
+export { TokenLabel } from "./token-label";
