+++ src/shared/ui/components/tooltip/index.ts
@@ -1,0 +1,1 @@
+export { Tooltip } from "./tooltip";
