+++ src/widgets/builder/index.ts
@@ -1,0 +1,1 @@
+export { Builder } from "./builder";
