+++ src/widgets/editor/index.ts
@@ -1,0 +1,2 @@
+export { Editor } from "./editor";
+export { TextInput, TextInputWithUnits } from "./elements"; // gone soon
