+++ src/widgets/menu/index.ts
@@ -1,0 +1,1 @@
+export { Menu } from "./menu";
