+++ src/widgets/sidebar/index.ts
@@ -1,0 +1,1 @@
+export { Sidebar } from "./sidebar";
